requirements.txt
Flask==2.0.1
requests==2.25.1
openai==0.11.3
pytest==6.2.4
pytest-cov==2.12.1
python-dotenv==0.17.1